+++ src/GapicClientTrait.php
@@ -128,7 +128,7 @@ trait GapicClientTrait
         }
     }
 
-    private static function initGrpcGcpConfig($hostName, $confPath)
+    private static function initGrpcGcpConfig(string $hostName, string $confPath)
     {
         $apiConfig = new ApiConfig();
         $apiConfig->mergeFromJsonString(file_get_contents($confPath));
@@ -255,7 +255,7 @@ trait GapicClientTrait
         return $options;
     }
 
-    private function shouldUseMtlsEndpoint($options)
+    private function shouldUseMtlsEndpoint(array $options)
     {
         $mtlsEndpointEnvVar = getenv('GOOGLE_API_USE_MTLS_ENDPOINT');
         if ('always' === $mtlsEndpointEnvVar) {
@@ -268,7 +268,7 @@ trait GapicClientTrait
         return !empty($options['clientCertSource']);
     }
 
-    private static function determineMtlsEndpoint($apiEndpoint)
+    private static function determineMtlsEndpoint(string $apiEndpoint)
     {
         $parts = explode('.', $apiEndpoint);
         if (count($parts) < 3) {
@@ -451,7 +451,7 @@ trait GapicClientTrait
      * @throws ValidationException
      */
     private function createTransport(
-        $apiEndpoint,
+        string $apiEndpoint,
         $transport,
         array $transportConfig,
         callable $clientCertSource = null
@@ -530,6 +530,94 @@ trait GapicClientTrait
 
     /**
      * @param string $methodName
+     * @param string $interfaceName
+     * @param Message $request
+     * @param array $optionalArgs {
+     *     Call Options
+     *
+     *     @type array $headers [optional] key-value array containing headers
+     *     @type int $timeoutMillis [optional] the timeout in milliseconds for the call
+     *     @type array $transportOptions [optional] transport-specific call options
+     *     @type RetrySettings|array $retrySettings [optional] A retry settings
+     *           override for the call.
+     * }
+     *
+     * @experimental
+     *
+     * @return PromiseInterface|PagedListResponse|BidiStream|ClientStream|ServerStream
+     */
+    private function startApiCall(
+        string $methodName,
+        string $interfaceName = null,
+        Message $request = null,
+        array $optionalArgs = []
+    ) {
+        // Ensure a method descriptor exists for the target method.
+        if (!isset($this->descriptors[$methodName])) {
+            throw new ValidationException("Requested method '$methodName' does not exist in descriptor configuration.");
+        }
+        $method = $this->descriptors[$methodName];
+        
+        // Prepare request-based headers, merge with user-provided headers,
+        // which take precedence.
+        $headerParams = $method['headerParams'] ?? [];
+        $requestHeaders = $this->buildRequestParamsHeader($headerParams, $request);
+        $optionalArgs['headers'] = array_merge($requestHeaders, $optionalArgs['headers'] ?? []);
+
+        // Default the interface name, if not set, to the client's protobuf service name.
+        $interfaceName = $interfaceName ?: $this->serviceName;
+
+        // Handle call based on call type configured in the method descriptor config.
+        if (!isset($method['callType'])) {
+            throw new ValidationException("Requested method '$methodName' does not have a callType " .
+                "in descriptor configuration.");
+        }
+        $callType = $method['callType'];
+        
+        if ($callType == Call::LONGRUNNING_CALL) {
+            if (!isset($method['longRunning'])) {
+                throw new ValidationException("Requested method '$methodName' does not have a longRunning config " .
+                    "in descriptor configuration.");
+            }
+            // @TODO: check if the client implements `OperationsClientInterface` instead
+            if (!method_exists($this, 'getOperationsClient')) {
+                throw new ValidationException("Client missing required getOperationsClient " .
+                    "for longrunning call '$methodName'");
+            }
+            return $this->startOperationsCall(
+                $methodName,
+                $optionalArgs,
+                $request,
+                $this->getOperationsClient(),
+                $interfaceName,
+                // Custom operations will define their own operation response type, whereas standard
+                // LRO defaults to the same type.
+                $method['responseType'] ?? null
+            );
+        }
+
+        // Fully-qualified name of the response message PHP class.
+        if (!isset($method['responseType'])) {
+            throw new ValidationException("Requested method '$methodName' does not have a responseType " .
+                "in descriptor configuration.");
+        }
+        $decodeType = $method['responseType'];
+
+        if ($callType == Call::PAGINATED_CALL) {
+            if (!isset($method['pageStreaming'])) {
+                throw new ValidationException("Requested method '$methodName' with callType PAGINATED_CALL does not " .
+                    "have a pageStreaming in descriptor configuration.");
+            }
+
+            return $this->getPagedListResponse($methodName, $optionalArgs, $decodeType, $request, $interfaceName);
+        }
+
+        // Unary, and all Streaming types handled by startCall.
+        return $this->startCall($methodName, $decodeType, $optionalArgs, $request, $callType, $interfaceName);
+    }
+
+    /**
+     * @param string $methodName
      * @param string $decodeType
      * @param array $optionalArgs {
      *     Call Options
@@ -547,12 +635,12 @@ trait GapicClientTrait
      * @return PromiseInterface|BidiStream|ClientStream|ServerStream
      */
     private function startCall(
-        $methodName,
-        $decodeType,
+        string $methodName,
+        string $decodeType,
         array $optionalArgs = [],
         Message $request = null,
-        $callType = Call::UNARY_CALL,
-        $interfaceName = null
+        int $callType = Call::UNARY_CALL,
+        string $interfaceName = null
     ) {
         $callStack = $this->createCallStack(
             $this->configureCallConstructionOptions($methodName, $optionalArgs)
@@ -633,7 +721,7 @@ trait GapicClientTrait
      *
      * @return array
      */
-    private function configureCallConstructionOptions($methodName, array $optionalArgs)
+    private function configureCallConstructionOptions(string $methodName, array $optionalArgs)
     {
         $retrySettings = $this->retrySettings[$methodName];
         // Allow for retry settings to be changed at call time
@@ -669,12 +757,12 @@ trait GapicClientTrait
      * @return PromiseInterface
      */
     private function startOperationsCall(
-        $methodName,
+        string $methodName,
         array $optionalArgs,
         Message $request,
         $client,
-        $interfaceName = null,
-        $operationClass = null
+        string $interfaceName = null,
+        string $operationClass = null
     ) {
         $callStack = $this->createCallStack(
             $this->configureCallConstructionOptions($methodName, $optionalArgs)
@@ -719,11 +807,11 @@ trait GapicClientTrait
      * @return PagedListResponse
      */
     private function getPagedListResponse(
-        $methodName,
+        string $methodName,
         array $optionalArgs,
-        $decodeType,
+        string $decodeType,
         Message $request,
-        $interfaceName = null
+        string $interfaceName = null
     ) {
         $callStack = $this->createCallStack(
             $this->configureCallConstructionOptions($methodName, $optionalArgs)
@@ -753,13 +841,72 @@ trait GapicClientTrait
      *
      * @return string
      */
-    private function buildMethod($interfaceName, $methodName)
+    private function buildMethod(string $interfaceName = null, string $methodName = null)
     {
         return sprintf(
             '%s/%s',
             $interfaceName ?: $this->serviceName,
             $methodName
         );
+    }
+
+    /**
+     * @param array $headerParams
+     * @param Message|null $request
+     *
+     * @return array
+     */
+    private function buildRequestParamsHeader(array $headerParams, Message $request = null)
+    {
+        $headers = [];
+        
+        // No request message means no request-based headers.
+        if (!$request) {
+            return $headers;
+        }
+        
+        foreach ($headerParams as $headerParam) {
+            $msg = $request;
+            $value = null;
+            foreach ($headerParam['fieldAccessors'] as $accessor) {
+                $value = $msg->$accessor();
+                
+                // In case the field in question is nested in another message,
+                // skip the header param when the nested message field is unset.
+                $msg = $value;
+                if (is_null($msg)) {
+                    break;
+                }
+            }
+
+            $keyName = $headerParam['keyName'];
+            
+            // If there are value pattern matchers configured and the target
+            // field was set, evaluate the matchers in the order that they were
+            // annotated in with last one matching wins.
+            $original = $value;
+            $matchers = isset($headerParam['matchers']) && !is_null($value) ?
+                $headerParam['matchers'] :
+                [];
+            foreach ($matchers as $matcher) {
+                $matches = [];
+                if (preg_match($matcher, $original, $matches)) {
+                    $value = $matches[$keyName];
+                }
+            }
+
+            // If there are no matches or the target field was unset, skip this
+            // header param.
+            if (!$value) {
+                continue;
+            }
+
+            $headers[$keyName] = $value;
+        }
+
+        $requestParams = new RequestParamsHeaderDescriptor($headers);
+
+        return $requestParams->getHeader();
     }
 
     /**